feature: refactor file-fetcher module
