+++ pom.xml
@@ -104,6 +104,7 @@
 
         <module>scaleph-dist</module>
         <module>scaleph-sql-gateway</module>
+        <module>scaleph-file-fetcher</module>
 
     </modules>
