+++ scaleph-file-fetcher/src/main/java/cn/sliew/scaleph/file/fetcher/cli/CliArgsException.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cn.sliew.scaleph.file.fetcher.cli;
+
+import cn.sliew.scaleph.common.exception.ScalephException;
+
+/**
+ * Special exception that is thrown when the command line parsing fails.
+ */
+public class CliArgsException extends ScalephException {
+
+    private static final long serialVersionUID = 1L;
+
+    public CliArgsException(String message) {
+        super(message);
+    }
+
+    public CliArgsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
