+++ scaleph-file-fetcher/src/main/java/cn/sliew/scaleph/file/fetcher/internal/InternalFileFetcherFactory.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cn.sliew.scaleph.file.fetcher.internal;
+
+import cn.sliew.scaleph.file.fetcher.FileFetcher;
+import cn.sliew.scaleph.file.fetcher.FileFetcherFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.net.URI;
+import java.util.Optional;
+import java.util.Properties;
+
+@Component
+public class InternalFileFetcherFactory implements FileFetcherFactory {
+
+    @Autowired
+    private InternalFileFetcher internalFileFetcher;
+
+    @Override
+    public Optional<FileFetcher> find(URI uri, Properties props) {
+        return Optional.of(internalFileFetcher);
+    }
+}
