+++ scaleph-file-fetcher/src/main/resources/application-dev.yml
@@ -1,0 +1,42 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+file-system:
+  #  type: local
+  type: s3
+  proxy: false
+  bucket: ${spring.application.name}
+  endPoint: http://127.0.0.1:9000
+  accessKey: admin
+  secretKey: password
+#  type: oss
+#  bucket: ${spring.application.name}
+#  endPoint: Aliyun OSS endpoint to connect to
+#  accessKey: Aliyun access key ID
+#  secretKey: Aliyun access key secret
+#  type: hdfs
+#  hadoopConfPath: Hadoop conf path
+#  defaultFS: hdfs://localhost:9000
+
+# log
+logging:
+  level:
+    root: info
+    springfox.documentation: warn
+    org.springframework.boot.autoconfigure: warn
+    org.springframework.beans.factory: warn
+    org.springframework.context: warn
