--- scaleph-plugins/scaleph-plugin-file-fetcher/scaleph-plugin-file-fetcher-internal/src/main/java/cn/sliew/scaleph/CliEntrypoint.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cn.sliew.scaleph;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CliEntrypoint {
-
-    public static void main(String[] args) {
-        String[] customArgs = new String[]{
-                "-uri", "scaleph://scaleph/user/wangqi/jar/mysql/flink-csv-1.16.1.jar",
-                "-path", "/Users/wangqi/Documents/repository/sliew/scaleph/temp/flink-csv-1.16.1.jar",
-        };
-        SpringApplication.run(CliEntrypoint.class, customArgs);
-    }
-}
